Fix tests

+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_external_local_field.py
@@ -16,7 +16,7 @@ import numpy as np
 import pytest
 
 import gt4py.next as gtx
-from gt4py.next import int64, neighbor_sum
+from gt4py.next import int32, neighbor_sum
 from gt4py.next.program_processors.runners import gtfn_cpu
 
 from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils import (
@@ -29,13 +29,13 @@ from next_tests.integration_tests.feature_tests.ffront_tests.ffront_test_utils i
 
 def test_external_local_field(reduction_setup, fieldview_backend):
     V2EDim, V2E = reduction_setup.V2EDim, reduction_setup.V2E
-    inp = gtx.np_as_located_field(Vertex, V2EDim)(reduction_setup.v2e_table)
-    ones = gtx.np_as_located_field(Edge)(np.ones(reduction_setup.num_edges, dtype=int64))
+    inp = gtx.np_as_located_field(Vertex, V2EDim)(reduction_setup.v2e_table, dtype=int32)
+    ones = gtx.np_as_located_field(Edge)(np.ones(reduction_setup.num_edges, dtype=int32))
 
     @gtx.field_operator(backend=fieldview_backend)
     def testee(
-        inp: gtx.Field[[Vertex, V2EDim], int64], ones: gtx.Field[[Edge], int64]
-    ) -> gtx.Field[[Vertex], int64]:
+        inp: gtx.Field[[Vertex, V2EDim], int32], ones: gtx.Field[[Edge], int32]
+    ) -> gtx.Field[[Vertex], int32]:
         return neighbor_sum(
             inp * ones(V2E), axis=V2EDim
         )  # multiplication with shifted `ones` because reduction of only non-shifted field with local dimension is not supported
@@ -53,10 +53,10 @@ def test_external_local_field_only(reduction_setup, fieldview_backend):
         )
 
     V2EDim = reduction_setup.V2EDim
-    inp = gtx.np_as_located_field(Vertex, V2EDim)(reduction_setup.v2e_table)
+    inp = gtx.np_as_located_field(Vertex, V2EDim)(reduction_setup.v2e_table, dtype=int32)
 
     @gtx.field_operator(backend=fieldview_backend)
-    def testee(inp: gtx.Field[[Vertex, V2EDim], int64]) -> gtx.Field[[Vertex], int64]:
+    def testee(inp: gtx.Field[[Vertex, V2EDim], int32]) -> gtx.Field[[Vertex], int32]:
         return neighbor_sum(inp, axis=V2EDim)
 
     testee(inp, out=reduction_setup.out, offset_provider=reduction_setup.offset_provider)
